agribot_index/requirements.txt
# Minio对象存储客户端
minio==7.2.16

# Milvus向量数据库客户端
pymilvus==2.5.15

# OpenAI API客户端（用于embedding）
openai==1.101.0

# 环境变量管理
python-dotenv==1.1.1

# PDF文档解析
pypdf2==3.0.1

# Office文档解析
markitdown[pdf, docx, pptx, xlsx, xls]==0.1.2

# text splitters
langchain-text-splitters==0.3.9

# PDF 渲染（用于扫描件OCR）
pymupdf

# MinerU文档解析（安装包非常大默认不安装，需要时再安装）
#mineru[core]==2.0.6

# Neo4j 知识图谱驱动
neo4j==5.27.0
